requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
appnope==0.1.3
asttokens==2.2.1
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
decorator==5.1.1
executing==1.2.0
FinMind==1.6.1
Flask==2.3.2
Flask-Cors==4.0.0
frozenlist==1.4.0
idna==3.4
importlib-metadata==6.8.0
iniconfig==2.0.0
ipython==8.14.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
loguru==0.7.0
lxml==4.9.3
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
multidict==6.0.4
numpy==1.25.2
packaging==23.1
pandas==2.0.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pluggy==1.2.0
prettytable==3.8.0
prompt-toolkit==3.0.39
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.1.1
pydantic_core==2.4.0
pyecharts==2.0.3
Pygments==2.16.1
pytest==7.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
simplejson==3.19.1
six==1.16.0
stack-data==0.6.2
ta==0.5.25
traitlets==5.9.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
wcwidth==0.2.6
Werkzeug==2.3.6
yarl==1.9.2
zipp==3.16.2
